driver: src: Fixed MisraC Violations
This patch fix the misra c 10.1 10.3 15.6 violations

+++ driver/src/device/xaie_device_aie.c
@@ -533,7 +533,7 @@ AieRC _XAie_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 	u8 flag;
 	u32 StartBit, EndBit;
 
-	u32 PartEndCol = DevInst->StartCol + DevInst->NumCols - 1;
+	u32 PartEndCol = (u32)(DevInst->StartCol + DevInst->NumCols - 1U);
 
 	if((Args->StartCol < DevInst->StartCol) || (Args->StartCol > PartEndCol) ||
 			((Args->StartCol + Args->NumCols - 1) > PartEndCol) ) {
@@ -544,7 +544,7 @@ AieRC _XAie_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 	/*Enable the clock control register for shims*/
 	for(u32 C = Args->StartCol; C < (Args->StartCol + Args->NumCols); C++) {
 
-		XAie_LocType TileLoc = XAie_TileLoc(C, 0);
+		XAie_LocType TileLoc = XAie_TileLoc((u8)C, 0U);
 
 		RC = _XAie_PmSetColumnClockBuffer(DevInst, TileLoc,
 				Args->Enable);
@@ -555,8 +555,10 @@ AieRC _XAie_SetColumnClk(XAie_DevInst *DevInst, XAie_BackendColumnReq *Args)
 		}
 	}
 
-	StartBit = _XAie_GetTileBitPosFromLoc(DevInst, XAie_TileLoc(Args->StartCol, 0));
-	EndBit = _XAie_GetTileBitPosFromLoc(DevInst, XAie_TileLoc(Args->StartCol + Args->NumCols, 0));
+	StartBit = _XAie_GetTileBitPosFromLoc(DevInst,
+			 XAie_TileLoc((u8)Args->StartCol, 0));
+	EndBit = _XAie_GetTileBitPosFromLoc(DevInst,
+			 XAie_TileLoc((u8)(Args->StartCol + Args->NumCols), 0));
 
 	if(Args->Enable) {
 		/*
